+ + move log to lib
