+++ app/lib.py
@@ -1,42 +1,57 @@
 import os
+import sys
+import logging
 import subprocess
+from pathlib import Path
+from sys import platform
+
+sys.path.append(str(Path(sys.argv[0]).absolute().parent.parent))
+from . import var
 
 
-# Generators for vars
-def get_base_dir():
-    base = os.path.join(os.getcwd())  # .. means parent directory
+# System log
+# Detect system/platform
+if platform == "linux" or platform == "linux2":
+    SYSTEM_LOG = '/var/log/ip2drop.log'
+elif platform == "darwin":
+    SYSTEM_LOG = os.path.join(var.EXPORTED_LOGS_DIR, 'ip2drop-script.log')
+elif platform == "win32":
+    print('Platform not supported. Exit. Bye.')
+    exit(1)
 
-    # Return the absolute path of the parent directory
-    return os.path.abspath(base)
-
-
-def up(n, nth_dir=os.getcwd()):
-    while n != 0:
-        nth_dir = os.path.dirname(nth_dir)
-        n -= 1
-    print(nth_dir)
-    return nth_dir
-
-
-def get_current_dir():
-    return os.path.dirname(os.path.abspath(__file__))
+# Logging
+logging.basicConfig(filename=SYSTEM_LOG,
+                    filemode='a',
+                    format='%(asctime)s,%(msecs)d %(name)s %(levelname)s %(message)s',
+                    datefmt='%d-%m-%Y %H-%M-%S',
+                    level=logging.DEBUG)
 
 
-def get_up_dir():
-    return os.path.dirname(get_current_dir())
+# Logger messages
+# TODO: Add -v, --verbose as DEBUG mode
+def log_debug(msg):
+    logging.debug(msg)
 
 
-def get_script_dir():
-    current = get_current_dir()
-    up_dir = os.path.dirname(current)
-    return up_dir
+def log_info(msg):
+    logging.info(msg)
 
 
-def get_parent_directory():
-    # Create a relative path to the parent of the current working directory
-    relative_parent = os.path.join(os.getcwd(), "..")  # .. means parent directory
-    # Return the absolute path of the parent directory
-    return os.path.abspath(relative_parent)
+def log_warn(msg):
+    logging.warning(msg)
+
+
+def log_err(msg):
+    logging.error(msg)
+
+
+def log_crit(msg):
+    logging.critical(msg)
+
+
+def msg_info(msg):
+    log_info(msg)
+    print(msg)
 
 
 # Command operators
@@ -55,11 +70,16 @@ def bash_cmd(cmd):
 # Ref: https://stackoverflow.com/questions/37487758/how-to-add-an-id-to-filename-before-extension
 def append_id(filename):
     name, ext = os.path.splitext(filename)
-    result = "{name}_{uid}{ext}".format(name=name, uid=TODAY.strftime("%d_%m_%Y"), ext=ext)
+    result = "{name}_{uid}{ext}".format(name=name, uid=var.TODAY.strftime("%d_%m_%Y"), ext=ext)
     # msg_info(f'Result: {result}')
     return result
 
-#
+def check_dir(dest):
+    is_exist = os.path.exists(dest)
+    if not is_exist:
+        # Create a new directory because it does not exist
+        os.makedirs(dest)
+        msg_info(f'Log catalog: {dest} created. Done.')
 
 def increment(number):
     number += 1
